+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 * [PR 422:](https://github.com/corretto/amazon-corretto-crypto-provider/pull/422) Support "pure" ML-DSA, Bump AWS-LC version to 1.43.0
 * [PR 423:](https://github.com/corretto/amazon-corretto-crypto-provider/pull/423) Support "External Mu" ML-DSA
 * [PR 424:](https://github.com/corretto/amazon-corretto-crypto-provider/pull/424) Use ACCP for ML-DSA in EvpKeyFactoryTest (for now)
+* [PR 426:](https://github.com/corretto/amazon-corretto-crypto-provider/pull/426) Add null check to AesCbcSpi
+* [PR 427:](https://github.com/corretto/amazon-corretto-crypto-provider/pull/427) Add provider info string
+* [PR 432:](https://github.com/corretto/amazon-corretto-crypto-provider/pull/432) Support Ed25519ph, bump AWS-LC to v1.46.0yy
 
 ## 2.4.1